+++ skema/skema-rs/mathml/src/lib.rs
@@ -5,3 +5,5 @@ pub mod expression;
 pub mod graph;
 pub mod normalization;
 pub mod parsing;
+pub mod mml2pn;
+pub mod petri_net;
